[Add] hideLayers
- hideLayers
    - 외부 클릭시 모든 레이어가 닫히는 이벤트
    - calendar의 경우, 외부 클릭 했을때에도 선택된 날짜와 연결선이 유지되는 문제가 발생
    - 이를 위해 외부 클릭시 걸려있던 startDate, endDate를 초기화시켜줌
    - 외부 클릭하면 startDate, endDate 선택이 모두 지워지도록 수정
