+++ src/calendars/functions/calendarDateStates.js
@@ -28,4 +28,12 @@ const removeState = (targetDate) => {
   selectedDateState[targetDate].element = null;
 };
 
-export { selectedDateState, createState, updateState, removeState };
+const initSelectedDateState = () => {
+  // calendar layer가 닫힐 때 값을 모두 초기화 시켜주는 함수
+  selectedDateState.startDate.data = [0, 0, 0];
+  selectedDateState.startDate.element = null;
+  selectedDateState.endDate.data = [0, 0, 0];
+  selectedDateState.endDate.element = null;
+};
+
+export { selectedDateState, createState, updateState, removeState, initSelectedDateState };
