+++ src/main.js
@@ -1,5 +1,6 @@
 import Toggle from "./toggle.js";
-import NavTab from "./nav.js";
+import NavTab from "./navTab/nav.js";
+import { hideLayers } from "./hideLayers.js";
 import { calendarCheckIn, checkInButton } from "./calendars/checkInCalendar.js";
 import { calendarCheckOut, checkOutButton } from "./calendars/checkOutCalendar.js";
 import { calendarExperienceDate, experienceDateButton, experienceDateStart, experienceDateEnd, placeholder } from "./calendars/experienceDateCalendar.js";
@@ -10,43 +11,45 @@ stay.eventListen();
 const experiences = new NavTab("experiences", "stay", "search-bar__experiences", "search-bar__stay");
 experiences.eventListen();
 
+hideLayers();
+
 // 계정 버튼
-const accountButton = document.querySelector("#account");
-const accountLayer = new Toggle("account__layer");
+// const accountButton = document.querySelector("#account");
+// const accountLayer = new Toggle("account__layer");
 
-accountButton.addEventListener("click", () => {
-  accountLayer.init();
-});
+// accountButton.addEventListener("click", () => {
+//   accountLayer.init();
+// });
 
-const isContain = (node, target) => node.contains(target);
+// const isContain = (node, target) => node.contains(target);
 
-document.addEventListener("click", ({ target }) => {
-  // 외부 클릭 시 레이어 숨기는 이벤트
+// document.addEventListener("click", ({ target }) => {
+//   // 외부 클릭 시 레이어 숨기는 이벤트
 
-  if (!isContain(accountLayer.element, target) && !isContain(accountButton, target)) {
-    // 계정 버튼 클릭시 나오는 레이어의 외부를 클릭했을 때
+//   if (!isContain(accountLayer.element, target) && !isContain(accountButton, target)) {
+//     // 계정 버튼 클릭시 나오는 레이어의 외부를 클릭했을 때
 
-    if (!isContain(checkInButton, target) && !isContain(calendarCheckIn.element, target)) {
-      // 체크인 버튼과 체크인 달력의 외부를 클릭했을 때
-      //   console.log("outside of calendarCheckIn");
-      calendarCheckIn.hide();
-      checkInButton.classList.remove("checked");
-    }
-    if (!isContain(checkOutButton, target) && !isContain(calendarCheckOut.element, target)) {
-      // 체크아웃 버튼과 체크아웃 달력의 외부를 클릭했을 때
-      //   console.log("outside of calendarCheckOut");
-      calendarCheckOut.hide();
-      checkOutButton.classList.remove("checked");
-    }
-    if (!isContain(experienceDateButton, target) && !isContain(calendarExperienceDate.element, target)) {
-      //   console.log(`outside of calendarExperienceDate`);
-      calendarExperienceDate.hide();
-      placeholder.classList.remove("none"); // 외부클릭 했으니까 원래 placeholder 글자 나오도록
-      experienceDateStart.innerText = ``;
-      experienceDateEnd.innerText = ``;
-      experienceDateButton.classList.remove("checked");
-    }
+//     if (!isContain(checkInButton, target) && !isContain(calendarCheckIn.element, target)) {
+//       // 체크인 버튼과 체크인 달력의 외부를 클릭했을 때
+//       //   console.log("outside of calendarCheckIn");
+//       calendarCheckIn.hide();
+//       checkInButton.classList.remove("checked");
+//     }
+//     if (!isContain(checkOutButton, target) && !isContain(calendarCheckOut.element, target)) {
+//       // 체크아웃 버튼과 체크아웃 달력의 외부를 클릭했을 때
+//       //   console.log("outside of calendarCheckOut");
+//       calendarCheckOut.hide();
+//       checkOutButton.classList.remove("checked");
+//     }
+//     if (!isContain(experienceDateButton, target) && !isContain(calendarExperienceDate.element, target)) {
+//       //   console.log(`outside of calendarExperienceDate`);
+//       calendarExperienceDate.hide();
+//       placeholder.classList.remove("none"); // 외부클릭 했으니까 원래 placeholder 글자 나오도록
+//       experienceDateStart.innerText = ``;
+//       experienceDateEnd.innerText = ``;
+//       experienceDateButton.classList.remove("checked");
+//     }
 
-    accountLayer.hide();
-  }
-});
+//     accountLayer.hide();
+//   }
+// });
